requirements.txt
opencv-python
PyYAML
face-recognition
ultralytics
numpy
# Note: hailo_platform is required for the Hailo accelerator.
# It is typically installed via the Hailo software suite or a specific .whl file provided by Hailo,
# so it is not listed here as a standard PyPI package.
